+++ clash-royale/clash_royale/__init__.py
@@ -1,7 +1,7 @@
 from gymnasium.envs.registration import register
 
 register(
-     id="clash_royale/Arena-v0",
-     entry_point="clash_royale.envs:ArenaEnv",
+     id="clash_royale/ClashRoyale",
+     entry_point="clash_royale.envs:ClashRoyaleEnv",
      max_episode_steps=3000,
 )
